new NodeProxy role 'setp'
A NodeProxy role that allows manipulating grains generated by a pattern
set as source of a NodeProxy. So far this seems to work for Pdefs set as
source, single or multi channel. If the pattern creates multichannel
output, a role for each channel has to be assigned and a key
'channelOffset' must be added to the Pbind in the role, denoting the
channel offset (relative to the Ndef's output channel).

Signed-off-by: Stefan Nussbaumer <st9fan@gmail.com>