NodeProxy:seti - allow arrays as input
If index or/and value are given as arrays seti should correctly replace
values at given indices. If indices exceed the size of values for the
given control name a valid index will always be returned through wrapAt.
If the values array is shorter than the indices array, the array will
wrap in order to always set an appropriate value.

Signed-off-by: Stefan Nussbaumer <st9fan@gmail.com>